Move calibration folder to library
