+++ docs/apidocs/calibration.rst
@@ -1,6 +1,6 @@
 .. _qiskit-experiments-calibration:
 
-.. automodule:: qiskit_experiments.calibration
+.. automodule:: qiskit_experiments.library.calibration
    :no-members:
    :no-inherited-members:
    :no-special-members:
