+++ qiskit_experiments/library/calibration/fine_amplitude.py
@@ -23,7 +23,9 @@ from qiskit.providers.options import Options
 from qiskit.pulse.schedule import ScheduleBlock
 
 from qiskit_experiments.base_experiment import BaseExperiment
-from qiskit_experiments.calibration.analysis.fine_amplitude_analysis import FineAmplitudeAnalysis
+from qiskit_experiments.library.calibration.analysis.fine_amplitude_analysis import (
+    FineAmplitudeAnalysis,
+)
 from qiskit_experiments.exceptions import CalibrationError
