+++ qiskit_experiments/library/calibration/rabi.py
@@ -23,7 +23,7 @@ import qiskit.pulse as pulse
 from qiskit.providers.options import Options
 
 from qiskit_experiments.base_experiment import BaseExperiment
-from qiskit_experiments.calibration.analysis.oscillation_analysis import OscillationAnalysis
+from qiskit_experiments.library.calibration.analysis.oscillation_analysis import OscillationAnalysis
 from qiskit_experiments.exceptions import CalibrationError
