+++ test/calibration/experiments/test_rabi.py
@@ -26,7 +26,7 @@ import qiskit.pulse as pulse
 from qiskit_experiments import ExperimentData
 from qiskit_experiments.library import Rabi, EFRabi
 
-from qiskit_experiments.calibration.analysis.oscillation_analysis import OscillationAnalysis
+from qiskit_experiments.library.calibration.analysis.oscillation_analysis import OscillationAnalysis
 from qiskit_experiments.data_processing.data_processor import DataProcessor
 from qiskit_experiments.data_processing.nodes import Probability
 from qiskit_experiments.composite.parallel_experiment import ParallelExperiment
